requirements.txt
Flask~=3.0.3
openai~=1.59.9
python-dotenv~=1.0.1
pydantic~=2.10.4
Werkzeug~=3.1.3
SQLAlchemy~=2.0.36
alembic~=1.14.0